Crypto transaction update

+++ PortfolioService/Models/PortfolioViewModel.cs
@@ -8,15 +8,15 @@ namespace PortfolioService.Models
     public class PortfolioViewModel
     {
         public List<CryptocurrencyHolding> Holdings { get; set; }
-        public decimal TotalValue { get; set; }
-        public decimal TotalProfitOrLoss { get; set; }
+        public double TotalValue { get; set; }
+        public double TotalProfitOrLoss { get; set; }
     }
 
     public class CryptocurrencyHolding
     {
         public string Name { get; set; }
-        public decimal Amount { get; set; }
-        public decimal Value { get; set; }
-        public decimal ProfitOrLoss { get; set; }
+        public double Amount { get; set; }
+        public double Value { get; set; }
+        public double ProfitOrLoss { get; set; }
     }
 }
