+++ PortfolioService/Models/SetAlertViewModel.cs
@@ -8,7 +8,7 @@ namespace PortfolioService.Models
     public class SetAlertViewModel
     {
         public string CryptocurrencyName { get; set; }
-        public decimal AlertThreshold { get; set; }
+        public double AlertThreshold { get; set; }
         public string Email { get; set; }
     }
 }
